requirements.txt
mysql-connector-python~=8.0.31
selenium~=4.7.2
pandas~=1.5.3
SQLAlchemy~=1.4.45
mysqlclient~=2.1.1
requests~=2.28.2
webdriver-manager==3.8.5
scikit-learn~=1.2.1
numpy~=1.24.2
Flask==2.2.2
Flask-Cors==3.0.10